requirements.txt
babel >= 2.8.0
discord.py >= 1.3.4
emoji >= 0.5.4
python-dateutil >= 2.8.1
